Refactor App.tsx into MainMenu, GameTable, and CardComponent
Extracted the monolithic `App` component into smaller, functional components:
- `CardComponent`: Reusable card display logic.
- `MainMenu`: Encapsulates main menu and multiplayer setup.
- `GameTable`: Handles the main game view (hand, table, players).

This improves readability and maintainability of the codebase.
Verified with tests and visual inspection.
